테스트 파일 내 트랜잭션 클라이언트 사용 구문 추가
