+++ src/test/kotlin/com/mashup/dhc/domain/model/UserRepositoryTest.kt
@@ -20,32 +20,22 @@ class UserRepositoryTest : BaseMongoDBTest() {
     }
 
     @Test
-    fun `test insert and find user`() =
+    fun testInsertAndFindUser() {
         runBlocking {
-            // Given
-            val user =
-                User(
-                    id = null,
-                    gender = Gender.MALE,
-                    userToken = "test-token",
-                    birthDate =
-                        BirthDate(
-                            LocalDate(2000, 1, 1),
-                            CalendarType.SOLAR
-                        ),
-                    birthTime = null,
-                    longTermMission =
-                        Mission(
-                            id = null,
-                            category = MissionCategory.FOOD,
-                            difficulty = 3,
-                            title = "밥 먹기",
-                            type = MissionType.DAILY,
-                            cost = Money("10.00"),
-                            endDate = null
-                        ),
-                    todayDailyMissionList =
-                        listOf(
+            mongoClient.startSession().let { session ->
+                // Given
+                val user =
+                    User(
+                        id = null,
+                        gender = Gender.MALE,
+                        userToken = "test-token",
+                        birthDate =
+                            BirthDate(
+                                LocalDate(2000, 1, 1),
+                                CalendarType.SOLAR
+                            ),
+                        birthTime = null,
+                        longTermMission =
                             Mission(
                                 id = null,
                                 category = MissionCategory.FOOD,
@@ -54,58 +44,63 @@ class UserRepositoryTest : BaseMongoDBTest() {
                                 type = MissionType.DAILY,
                                 cost = Money("10.00"),
                                 endDate = null
+                            ),
+                        todayDailyMissionList =
+                            listOf(
+                                Mission(
+                                    id = null,
+                                    category = MissionCategory.FOOD,
+                                    difficulty = 3,
+                                    title = "밥 먹기",
+                                    type = MissionType.DAILY,
+                                    cost = Money("10.00"),
+                                    endDate = null
+                                )
+                            ),
+                        pastRoutineHistoryIds = emptyList(),
+                        preferredMissionCategoryList = listOf(MissionCategory.FOOD),
+                        currentAmulet =
+                            Amulet(
+                                totalPiece = 0,
+                                remainPiece = 0
                             )
-                        ),
-                    pastRoutineHistoryIds = emptyList(),
-                    preferredMissionCategoryList = listOf(MissionCategory.FOOD),
-                    currentAmulet =
-                        Amulet(
-                            totalPiece = 0,
-                            remainPiece = 0
-                        )
-                )
+                    )
 
-            // When
-            val insertedId = userRepository.insertOne(user)
-            assertNotNull(insertedId)
+                // When
+                val insertedId = userRepository.insertOne(user, session)
+                assertNotNull(insertedId)
 
-            val objectId = insertedId.asObjectId().value
-            val foundUser = userRepository.findById(objectId)
+                val objectId = insertedId.asObjectId().value
+                val foundUser = userRepository.findById(objectId)
 
-            // Then
-            assertNotNull(foundUser)
-            assertEquals(Gender.MALE, foundUser.gender)
-            assertEquals("test-token", foundUser.userToken)
-            assertEquals(LocalDate(2000, 1, 1), foundUser.birthDate.date)
-            assertEquals(CalendarType.SOLAR, foundUser.birthDate.calendarType)
-            assertEquals(MissionCategory.FOOD, foundUser.longTermMission!!.category)
-            assertEquals(1, foundUser.todayDailyMissionList.size)
-            assertEquals(MissionCategory.FOOD, foundUser.todayDailyMissionList[0].category)
+                // Then
+                assertNotNull(foundUser)
+                assertEquals(Gender.MALE, foundUser.gender)
+                assertEquals("test-token", foundUser.userToken)
+                assertEquals(LocalDate(2000, 1, 1), foundUser.birthDate.date)
+                assertEquals(CalendarType.SOLAR, foundUser.birthDate.calendarType)
+                assertEquals(MissionCategory.FOOD, foundUser.longTermMission!!.category)
+                assertEquals(1, foundUser.todayDailyMissionList.size)
+                assertEquals(MissionCategory.FOOD, foundUser.todayDailyMissionList[0].category)
+
+                session.close()
+            }
         }
+    }
 
     @Test
-    fun `test update user`() =
+    fun testUpdateUser() {
         runBlocking {
-            // Given
-            val user =
-                User(
-                    id = null,
-                    gender = Gender.MALE,
-                    userToken = "test-token",
-                    birthDate = BirthDate(LocalDate(2000, 1, 1), CalendarType.SOLAR),
-                    birthTime = null,
-                    longTermMission =
-                        Mission(
-                            id = null,
-                            category = MissionCategory.FOOD,
-                            difficulty = 3,
-                            type = MissionType.DAILY,
-                            cost = Money("10.00"),
-                            title = "밥 먹기",
-                            endDate = null
-                        ),
-                    todayDailyMissionList =
-                        listOf(
+            mongoClient.startSession().let { session ->
+                // Given
+                val user =
+                    User(
+                        id = null,
+                        gender = Gender.MALE,
+                        userToken = "test-token",
+                        birthDate = BirthDate(LocalDate(2000, 1, 1), CalendarType.SOLAR),
+                        birthTime = null,
+                        longTermMission =
                             Mission(
                                 id = null,
                                 category = MissionCategory.FOOD,
@@ -114,113 +109,133 @@ class UserRepositoryTest : BaseMongoDBTest() {
                                 cost = Money("10.00"),
                                 title = "밥 먹기",
                                 endDate = null
+                            ),
+                        todayDailyMissionList =
+                            listOf(
+                                Mission(
+                                    id = null,
+                                    category = MissionCategory.FOOD,
+                                    difficulty = 3,
+                                    type = MissionType.DAILY,
+                                    cost = Money("10.00"),
+                                    title = "밥 먹기",
+                                    endDate = null
+                                )
+                            ),
+                        pastRoutineHistoryIds = emptyList(),
+                        preferredMissionCategoryList = listOf(MissionCategory.FOOD),
+                        currentAmulet =
+                            Amulet(
+                                totalPiece = 0,
+                                remainPiece = 0
                             )
-                        ),
-                    pastRoutineHistoryIds = emptyList(),
-                    preferredMissionCategoryList = listOf(MissionCategory.FOOD),
-                    currentAmulet =
-                        Amulet(
-                            totalPiece = 0,
-                            remainPiece = 0
-                        )
-                )
+                    )
 
-            // When
-            val insertedId = userRepository.insertOne(user)
-            assertNotNull(insertedId)
-            val objectId = insertedId.asObjectId().value
+                // When
+                val insertedId = userRepository.insertOne(user, session)
+                assertNotNull(insertedId)
+                val objectId = insertedId.asObjectId().value
 
-            val updatedUser =
-                User(
-                    id = objectId,
-                    gender = Gender.MALE,
-                    userToken = "test-token",
-                    birthDate = BirthDate(LocalDate(2000, 1, 1), CalendarType.SOLAR),
-                    birthTime = null,
-                    longTermMission =
-                        Mission(
-                            id = null,
-                            category = MissionCategory.FOOD,
-                            difficulty = 3,
-                            type = MissionType.DAILY,
-                            cost = Money("10.00"),
-                            title = "밥 먹기",
-                            endDate = null
-                        ),
-                    todayDailyMissionList =
-                        listOf(
+                val updatedUser =
+                    User(
+                        id = objectId,
+                        gender = Gender.MALE,
+                        userToken = "test-token",
+                        birthDate = BirthDate(LocalDate(2000, 1, 1), CalendarType.SOLAR),
+                        birthTime = null,
+                        longTermMission =
                             Mission(
                                 id = null,
-                                category = MissionCategory.TRANSPORTATION, // Changed category
-                                difficulty = 5, // Changed difficulty
+                                category = MissionCategory.FOOD,
+                                difficulty = 3,
                                 type = MissionType.DAILY,
-                                cost = Money("15.00"), // Changed cost
+                                cost = Money("10.00"),
                                 title = "밥 먹기",
                                 endDate = null
+                            ),
+                        todayDailyMissionList =
+                            listOf(
+                                Mission(
+                                    id = null,
+                                    category = MissionCategory.TRANSPORTATION, // Changed category
+                                    difficulty = 5, // Changed difficulty
+                                    type = MissionType.DAILY,
+                                    cost = Money("15.00"), // Changed cost
+                                    title = "밥 먹기",
+                                    endDate = null
+                                )
+                            ),
+                        pastRoutineHistoryIds = listOf(ObjectId()), // Added mission history
+                        preferredMissionCategoryList = listOf(MissionCategory.FOOD),
+                        currentAmulet =
+                            Amulet(
+                                totalPiece = 0,
+                                remainPiece = 0
                             )
-                        ),
-                    pastRoutineHistoryIds = listOf(ObjectId()), // Added mission history
-                    preferredMissionCategoryList = listOf(MissionCategory.FOOD),
-                    currentAmulet =
-                        Amulet(
-                            totalPiece = 0,
-                            remainPiece = 0
-                        )
-                )
+                    )
 
-            val updateCount = userRepository.updateOne(objectId, updatedUser)
-            assertEquals(1, updateCount)
+                val updateCount = userRepository.updateOne(objectId, updatedUser, session)
+                assertEquals(1, updateCount)
 
-            // Then
-            val foundUser = userRepository.findById(objectId)
-            assertNotNull(foundUser)
-            assertEquals(MissionCategory.TRANSPORTATION, foundUser.todayDailyMissionList[0].category)
-            assertEquals(5, foundUser.todayDailyMissionList[0].difficulty)
-            assertEquals(Money("15.00"), foundUser.todayDailyMissionList[0].cost)
-            assertEquals(1, foundUser.pastRoutineHistoryIds.size)
+                // Then
+                val foundUser = userRepository.findById(objectId)
+                assertNotNull(foundUser)
+                assertEquals(MissionCategory.TRANSPORTATION, foundUser.todayDailyMissionList[0].category)
+                assertEquals(5, foundUser.todayDailyMissionList[0].difficulty)
+                assertEquals(Money("15.00"), foundUser.todayDailyMissionList[0].cost)
+                assertEquals(1, foundUser.pastRoutineHistoryIds.size)
+
+                session.close()
+            }
         }
+    }
 
     @Test
-    fun `test delete user`() =
+    fun testDeleteUser() {
         runBlocking {
-            // Given
-            val user =
-                User(
-                    id = null,
-                    gender = Gender.MALE,
-                    userToken = "test-token",
-                    birthDate = BirthDate(LocalDate(2000, 1, 1), CalendarType.SOLAR),
-                    birthTime = null,
-                    longTermMission =
-                        Mission(
-                            id = null,
-                            category = MissionCategory.FOOD,
-                            difficulty = 3,
-                            type = MissionType.DAILY,
-                            cost = Money("10.00"),
-                            title = "밥 먹기",
-                            endDate = null
-                        ),
-                    todayDailyMissionList = emptyList(),
-                    pastRoutineHistoryIds = emptyList(),
-                    preferredMissionCategoryList = listOf(MissionCategory.FOOD),
-                    currentAmulet =
-                        Amulet(
-                            totalPiece = 0,
-                            remainPiece = 0
-                        )
-                )
+            mongoClient.startSession().let { session ->
+                // Given
+                val user =
+                    User(
+                        id = null,
+                        gender = Gender.MALE,
+                        userToken = "test-token",
+                        birthDate = BirthDate(LocalDate(2000, 1, 1), CalendarType.SOLAR),
+                        birthTime = null,
+                        longTermMission =
+                            Mission(
+                                id = null,
+                                category = MissionCategory.FOOD,
+                                difficulty = 3,
+                                type = MissionType.DAILY,
+                                cost = Money("10.00"),
+                                title = "밥 먹기",
+                                endDate = null
+                            ),
+                        todayDailyMissionList = emptyList(),
+                        pastRoutineHistoryIds = emptyList(),
+                        preferredMissionCategoryList = listOf(MissionCategory.FOOD),
+                        currentAmulet =
+                            Amulet(
+                                totalPiece = 0,
+                                remainPiece = 0
+                            )
+                    )
 
-            // When
-            val insertedId = userRepository.insertOne(user)
-            assertNotNull(insertedId)
-            val objectId = insertedId.asObjectId().value
+                // When
+                val insertedId = userRepository.insertOne(user, session)
+                assertNotNull(insertedId)
+                val objectId = insertedId.asObjectId().value
 
-            val deleteCount = userRepository.deleteById(objectId)
-            assertEquals(1, deleteCount)
+                val deleteCount = userRepository.deleteById(objectId, session)
+                assertEquals(1, deleteCount)
 
-            // Then
-            val foundUser = userRepository.findById(objectId)
-            assertNull(foundUser)
+                // Then
+                val foundUser = userRepository.findById(objectId)
+                assertNull(foundUser)
+
+                session.close()
+            }
         }
+    }
 }
